requirements.txt
streamlit>=1.28.0
openai>=1.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
pendulum>=2.1.2
pydantic>=2.5.0
python-dotenv>=1.0.0

